refactor: consolidate external labs types and update routing
